% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\docType{package}
\name{commotions}
\alias{commotions}
\alias{commotions-package}
\title{A road-user model for commotions project}
\description{
A road-user model for commotions project
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com>, Gustav Markkula <G.Markkula@leeds.ac.uk>
}
\keyword{package}
