% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{find_collision_point}
\alias{find_collision_point}
\title{Find first possible collision point}
\usage{
find_collision_point(obj, collision_distance = 1)
}
\arguments{
\item{obj}{a list storing trajectory (x, y), speed, heading angles, and 
time stamps from a simulation run}

\item{collision_distance}{a small number considering a collision is imminent}
}
\value{
a list of three elements.
}
\description{
Test whether two line segments cross each other.
}
