% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_intersection_of_lines}
\alias{get_intersection_of_lines}
\title{get_intersection_of_lines}
\usage{
get_intersection_of_lines(line1_pointA, line1_pointB, line2_pointA,
  line2_pointB)
}
\arguments{
\item{line1_pointA}{self_pos}

\item{line1_pointB}{self_goal}

\item{line2_pointA}{other_pos}

\item{line2_pointB}{other_goal}
}
\value{
a vector
}
\description{
get_intersection_of_lines
}
\examples{
pos  <- matrix(c(0, -5, 40, 0), ncol=2); pos
goal <- matrix(c(0, 5, -50, 0), ncol=2); goal
self_pos <- pos[,1]
self_goal <- goal[,1] 
other_pos <- pos[,2]
other_goal <- goal[,2] 
res <- get_intersection_of_lines(self_pos, self_goal, other_pos, other_goal)
## [1] 0 0
}
