% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{is_collision}
\alias{is_collision}
\title{Is on a Collision Course}
\usage{
is_collision(obj)
}
\arguments{
\item{obj}{an output from test_Run}
}
\description{
Is on a Collision Course using the output from Simulation and Kinmatics 
class
}
