% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_cross}
\alias{is_cross}
\title{Is Two Segment Parallel?}
\usage{
is_cross(segment0, segment1)
}
\arguments{
\item{segment0}{the four coordinates of the first segment, c(x0, y0, x1, y1)}

\item{segment1}{the four coordinates of the second segment, c(x0, y0, x1, y1)}
}
\value{
Boolean TRUE or FALSE
}
\description{
Test whether two segments run parallel or across each other.
}
\examples{
pos <- matrix(c(-5, -5, -5, 5), ncol=2)
segment0 <- c(pos[1,], pos[2,])
wall <- c(-5, 0, -2.5, 0)
is_cross(segment0, wall)
}
\references{
https://math.stackexchange.com/questions/1342435/how-to-determine-if-2-line-segments-cross
}
