% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_2traj}
\alias{plot_2traj}
\title{Plot travelling trajectory}
\usage{
plot_2traj(obj, wall = NULL, fixed_x = FALSE, fixed_y = FALSE,
  collision_distance = 1, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}

\item{wall}{draw a wall segment}

\item{fixed_x}{x limit for the trajectory}

\item{fixed_y}{y limit for the trajectory}
}
\description{
Trajectory of two pedestrian agents using Simulation and Kinematics Classes
}
