% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_acc}
\alias{plot_acc}
\title{Plot acceleration trajectory mix case}
\usage{
plot_acc(obj, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}
}
\description{
Plot acceleration trajectory
}
