% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_speed}
\alias{plot_speed}
\title{Plot speed trajectory mix case}
\usage{
plot_speed(obj, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}
}
\description{
Plot speed trajectory
}
