% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_trajectory_AC}
\alias{plot_trajectory_AC}
\title{Plot travelling trajectory}
\usage{
plot_trajectory_AC(obj, wall = NULL, fixed_x = FALSE,
  fixed_y = FALSE, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}

\item{wall}{draw a wall segment}

\item{fixed_x}{x limit for the trajectory}

\item{fixed_y}{y limit for the trajectory}
}
\description{
Trajectory of two pedestrian agents using Agent Class
}
