% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trajOnly}
\alias{plot_trajOnly}
\title{Plot travelling trajectory}
\usage{
plot_trajOnly(obj, fixed_x = FALSE, fixed_y = FALSE, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}

\item{fixed_x}{x limit for the trajectory}

\item{fixed_y}{y limit for the trajectory}
}
\description{
Trajectory of two pedestrian agents using Simulation and Kinematics Classes
}
