% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_tsee}
\alias{plot_tsee}
\title{Plot Tsee trajectories}
\usage{
plot_tsee(obj, wall = NULL)
}
\arguments{
\item{obj}{a list storing trajectory (x, y), speed, heading angles, and 
time stamps from a simulation run}

\item{wall}{the wall segment}
}
\description{
Find first possible collision point
}
