% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_ttc}
\alias{plot_ttc}
\title{Plot time to collision trajectory one agent}
\usage{
plot_ttc(obj, fixed_x = FALSE, fixed_y = FALSE)
}
\description{
Trajectory of a pedestrian agent and a vehicle agent. Note when using
fixed_y, the first limit should set a small value, instead of 0.
}
