% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commotions.R
\name{plot_V2V_1Agent}
\alias{plot_V2V_1Agent}
\title{Plot travelling trajectory}
\usage{
plot_V2V_1Agent(obj, fixed_x = FALSE, fixed_y = FALSE, fn = NULL)
}
\arguments{
\item{obj}{an output from test_Run}

\item{fixed_x}{x limit for the trajectory}

\item{fixed_y}{y limit for the trajectory}

\item{wall}{draw a wall segment}
}
\description{
Trajectory of a pedestrian agent and a vehicle agent
}
