% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Simulate Either Walkers' and Vehicles' Trajectory}
\usage{
run(time, para, pos, goal, so, ao, ss, sa, obstacles, cd)
}
\description{
This function uses (new) Agent Class. Not yet fully tested
}
