% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_P2P}
\alias{test_P2P}
\title{Simulate P2P Trajectory}
\usage{
test_P2P(time, para, pos, goal, so, ao, ss, sa, obstacles, cd, sr, wall)
}
\description{
This function uses Simulation and Kinematics Classes
}
