% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_P2V}
\alias{test_P2V}
\title{Simulate P2V Trajectory}
\usage{
test_P2V(time, para, pos, goal, so, ao, ss, sa, obstacles, cd, sr)
}
\description{
This function uses Simulation and Kinematics Classes
}
